refactor(guards): the section slicer moves out of the claim guard
The identity guard needs the same cut, and two guards slicing differently would
judge two different documents. Behaviour-preserving: the 42 claim assertions are
green off the shared helper.
